refactor: extract cloud form into service
