+++ src/app/features/cloud/components/cloud-form-section.component.ts
@@ -1,11 +1,12 @@
 import { Component, computed, input, Signal, signal, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { costRanges, defaultValues, formContext } from '../../../carbon-estimator-form/carbon-estimator-form.constants';
+import { costRanges, formContext } from '../../../carbon-estimator-form/carbon-estimator-form.constants';
 import { FormatCostRangePipe } from '../../../pipes/format-cost-range.pipe';
 import { compareCostRanges } from '../../../utils/cost-range';
 import { SectionHeaderComponent } from '../../../components/section-header/section-header.component';
 import { LocationInputComponent } from '../../../components/location-input/location-input.component';
+import { defaultValues } from '../../../services/form.service';
 
 @Component({
   selector: 'app-cloud-form-section',
