+++ src/app/features/cloud/services/estimate-cloud-emissions.service.ts
@@ -1,9 +1,10 @@
 import { inject, Injectable } from '@angular/core';
-import { Cloud, CostRange } from '../../../types/carbon-estimator';
+import { CostRange } from '../../../types/carbon-estimator';
 import { gCo2ePerKwh, KgCo2e, KilowattHour } from '../../../types/units';
 import { CLOUD_AVERAGE_PUE } from '../../../estimation/constants';
 import { estimateEnergyEmissions } from '../../../estimation/estimate-energy-emissions';
 import { CarbonIntensityService } from '../../../services/carbon-intensity.service';
+import { Cloud } from './cloud-form.service';
 
 @Injectable({
   providedIn: 'root',
