+++ src/app/services/form.service.ts
@@ -1,8 +1,31 @@
 import { inject, Injectable } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { SaasFormService } from '../features/saas/services/saas-form.service';
+import { defaultSaasValues, SaasFormService } from '../features/saas/services/saas-form.service';
 import { EstimatorFormValues, EstimatorValues, WorldLocation } from '../types/carbon-estimator';
-import { defaultValues, EstimatorFormRawValue } from '../carbon-estimator-form/carbon-estimator-form.constants';
+import { EstimatorFormRawValue } from '../carbon-estimator-form/carbon-estimator-form.constants';
+import { CloudFormService, defaultCloudValues } from '../features/cloud/services/cloud-form.service';
+
+export const defaultValues: Required<EstimatorValues> = {
+  upstream: {
+    headCount: 100,
+    desktopPercentage: 50,
+    employeeLocation: 'WORLD',
+  },
+  onPremise: {
+    estimateServerCount: false,
+    serverLocation: 'WORLD',
+    numberOfServers: 10,
+  },
+  cloud: defaultCloudValues,
+  downstream: {
+    noDownstream: false,
+    customerLocation: 'WORLD',
+    monthlyActiveUsers: 100,
+    mobilePercentage: 50,
+    purposeOfSite: 'average',
+  },
+  saas: defaultSaasValues,
+};
 
 @Injectable({
   providedIn: 'root',
@@ -10,6 +33,7 @@ import { defaultValues, EstimatorFormRawValue } from '../carbon-estimator-form/c
 export class FormService {
   private formBuilder = inject(FormBuilder);
   private saasFormService = inject(SaasFormService);
+  private cloudFormService = inject(CloudFormService);
 
   estimatorForm: FormGroup<EstimatorFormValues> = this.initialise();
 
@@ -25,12 +49,7 @@ export class FormService {
         serverLocation: [defaultValues.onPremise.serverLocation as WorldLocation | 'unknown'],
         numberOfServers: [defaultValues.onPremise.numberOfServers, [Validators.required, Validators.min(0)]],
       }),
-      cloud: this.formBuilder.nonNullable.group({
-        noCloudServices: [false],
-        cloudLocation: [defaultValues.cloud.cloudLocation as WorldLocation | 'unknown'],
-        cloudPercentage: [defaultValues.cloud.cloudPercentage],
-        monthlyCloudBill: [defaultValues.cloud.monthlyCloudBill],
-      }),
+      cloud: this.cloudFormService.form,
       downstream: this.formBuilder.nonNullable.group({
         noDownstream: [false],
         customerLocation: [defaultValues.downstream.customerLocation],
